Changed the layout to many components, who were placed in a seperate components/navbar directory
