+++ src/lib/components/navbar/navbarLogo.svelte
@@ -1,0 +1,4 @@
+<div class="flex items-center space-x-2">
+    <img src="./src/assets/logo.png" alt="Logo" class="h-6" />
+    <span class="text-sm text-white font-semibold">FoodieFuse</span>
+</div>
