+++ src/lib/components/navbar/navbarProfileDisplay.svelte
